codebase_L13_CAPSTONE2025/codebase_L13_CAPSTONE2025/requirements.txt
# torch>=2.0.1
numpy
pandas
scipy
soundfile
ml_collections
tqdm
segmentation_models_pytorch==0.3.3
timm==0.9.2
audiomentations==0.24.0
pedalboard==0.8.1
omegaconf==2.2.3
beartype==0.14.1
rotary_embedding_torch==0.3.5
einops==0.6.1
librosa
demucs==4.0.0
transformers==4.35.0
torchmetrics==0.11.4
spafe==0.3.2
protobuf==3.20.3
torch_audiomentations
asteroid==0.7.0
auraloss
